Reorganized distributions::weighted module
